Split SwiftData schemas and add migration fixtures
