+++ Symi/Sources/Models/CurrentModelAliases.swift
@@ -1,0 +1,10 @@
+import Foundation
+import SwiftData
+
+typealias Episode = SymiSchemaV6.Episode
+typealias MedicationEntry = SymiSchemaV6.MedicationEntry
+typealias MedicationDefinition = SymiSchemaV6.MedicationDefinition
+typealias ContinuousMedication = SymiSchemaV6.ContinuousMedication
+typealias ContinuousMedicationCheck = SymiSchemaV6.ContinuousMedicationCheck
+typealias WeatherSnapshot = SymiSchemaV6.WeatherSnapshot
+// Doctor- und Terminmodelle wurden mit Schema V5 aus dem aktiven Datenmodell entfernt.
